+++ src/cli/app.rs
@@ -1,3 +1,5 @@
+use clap::{Parser, Subcommand};
+
 #[derive(Clone, Debug, Parser)]
 #[command(name="Pulith",version=env!("CARGO_PKG_VERSION"),about,long_about=None,propagate_version=true)]
 pub struct App {
