+++ src/main.rs
@@ -1,3 +1,6 @@
+mod backend;
+mod cli;
+mod reg;
 fn main() {
     println!("Hello, world!");
 }
